+++ src/writers/variant_categorizer.rs
@@ -1,6 +1,7 @@
 
 use anyhow::{Context, bail};
 use log::debug;
+use noodles::bgzf;
 use noodles::core::Position;
 use noodles::vcf;
 use noodles::vcf::header::record::value::{Map, map};
@@ -9,7 +10,8 @@ use noodles_util::variant::io::indexed_reader::Builder as VcfBuilder;
 use noodles::vcf::variant::record::samples::keys::key as vcf_key;
 use noodles::vcf::variant::record_buf;
 use std::collections::BTreeMap;
-use std::path::Path;
+use std::fs::File;
+use std::path::{Path, PathBuf};
 
 use crate::data_types::compare_benchmark::CompareBenchmark;
 use crate::data_types::compare_region::CompareRegion;
@@ -24,54 +26,41 @@ pub const REGION_ID: &str = "RI";
 
 /// Wrapper struct that will parse the results from the core processes and determine where to write variants
 pub struct VariantCategorizer {
-    /// Header that goes into the truth VCFs
-    truth_header: vcf::Header,
-    /// Header that goes into the query VCFs
-    query_header: vcf::Header,
-    /// Contains links to all of the various VCF writers that we have by their source; e.g. truth
-    vcf_writers: BTreeMap<VariantSource, vcf::io::Writer<Box<dyn std::io::Write>>>
+    /// Either truth or query
+    source: VariantSource,
+    /// Track the filename
+    filename: PathBuf,
+    /// Header that goes into the VCF
+    vcf_header: vcf::Header,
+    /// The actual writer
+    vcf_writer: vcf::io::Writer<bgzf::MultithreadedWriter<File>>
 }
 
 impl VariantCategorizer {
-    /// Contructor that loads the truth and query VCFs and create the files for writing.
+    /// Contructor that loads the truth and query VCFs and create a pair of writers for the outputs.
     /// # Arguments
     /// * `truth_vcf_fn` - the input Truth VCF
     /// * `truth_sample_name` - the sample name for the truth VCFs
     /// * `query_vcf_fn` - the input Query VCF
     /// * `query_sample_name` - the sample name for the query VCFs
     /// * `debug_folder` - root of the debug folder, this will create files inside it
+    /// * `threads` - writer threads per gzip writer
     /// # Errors
     /// * if there are any problems opening the input files or writing files to the debug folder
-    pub fn new(
+    pub fn new_writer_pair(
         truth_vcf_fn: &Path,
         truth_sample_name: String,
         query_vcf_fn: &Path,
         query_sample_name: String,
-        debug_folder: &Path
-    ) -> anyhow::Result<Self> {
-        // Open the VCF files
-        let mut truth_vcf_reader = VcfBuilder::default()
-            .build_from_path(truth_vcf_fn)
-            .with_context(|| format!("Error while opening {truth_vcf_fn:?}:"))?;
-        let mut query_vcf_reader = VcfBuilder::default()
-            .build_from_path(query_vcf_fn)
-            .with_context(|| format!("Error while opening {query_vcf_fn:?}:"))?;
+        debug_folder: &Path,
+        threads: usize
+    ) -> anyhow::Result<(Self, Self)> {
+        let mut source_lookup: BTreeMap<VariantSource, Self> = Default::default();
 
-        // get the headers also
-        let mut truth_vcf_header = truth_vcf_reader.read_header()
-            .with_context(|| format!("Error while reading header of {truth_vcf_fn:?}:"))?;
-        let mut query_vcf_header = query_vcf_reader.read_header()
-            .with_context(|| format!("Error while reading header of {query_vcf_fn:?}:"))?;
-
+        // constants we add
         let ver: &str = crate::cli::core::FULL_VERSION.as_str(); // clippy gets weird about direct access
         let cli_version = format!("\"{}\"", ver);
         let cli_string = format!("\"{}\"", std::env::args().collect::<Vec<String>>().join(" "));
-        truth_vcf_header.insert("aardvark_version".parse()?, vcf::header::record::Value::from(cli_version.clone()))?;
-        truth_vcf_header.insert("aardvark_command".parse()?, vcf::header::record::Value::from(cli_string.clone()))?;
-        query_vcf_header.insert("aardvark_version".parse()?, vcf::header::record::Value::from(cli_version))?;
-        query_vcf_header.insert("aardvark_command".parse()?, vcf::header::record::Value::from(cli_string))?;
-
-        // add expected and observed allele counts to our header
         let extra_header = [
             (
                 BENCHMARK_DECISION_KEY.to_string(),
@@ -90,51 +79,68 @@ impl VariantCategorizer {
                 Map::<map::Format>::new(map::format::Number::Count(1), map::format::Type::Integer, "Region ID for the comparison")
             )
         ];
-        
-        for (header_key, header_value) in extra_header.iter() {
-            truth_vcf_header.formats_mut().insert(header_key.clone(), header_value.clone());
-            query_vcf_header.formats_mut().insert(header_key.clone(), header_value.clone());
-        }
 
-        // set the truth and query to just have a single sample
-        let truth_samples = truth_vcf_header.sample_names_mut();
-        truth_samples.clear();
-        truth_samples.insert(truth_sample_name);
-
-        let query_samples = query_vcf_header.sample_names_mut();
-        query_samples.clear();
-        query_samples.insert(query_sample_name);
-
-        // we have a lot of sub-files to write
-        let filetypes = [
-            (VariantSource::Truth, "truth.vcf.gz"),
-            (VariantSource::Query, "query.vcf.gz"),
-        ];
-
-        let mut vcf_writers: BTreeMap<VariantSource, vcf::io::Writer<Box<dyn std::io::Write>>> = Default::default();
-        for (source, filename) in filetypes.into_iter() {
-            // make the VCF and write the header
-            let vcf_fn = debug_folder.join(filename);
-            debug!("Opening {vcf_fn:?} for writing...");
-            let mut vcf_writer = vcf::io::writer::Builder::default()
-                .set_compression_method(noodles::vcf::io::CompressionMethod::Bgzf)
-                .build_from_path(vcf_fn)?;
-
-            // write the header based on the data source
-            match source {
-                VariantSource::Truth => vcf_writer.write_header(&truth_vcf_header)?,
-                VariantSource::Query => vcf_writer.write_header(&query_vcf_header)?,
+        for variant_source in [VariantSource::Truth, VariantSource::Query] {
+            let input_fn = match variant_source {
+                VariantSource::Truth => truth_vcf_fn,
+                VariantSource::Query => query_vcf_fn
+            };
+            let sample_name = match variant_source {
+                VariantSource::Truth => truth_sample_name.clone(),
+                VariantSource::Query => query_sample_name.clone()
             };
 
+            // Open the VCF file
+            let mut vcf_reader = VcfBuilder::default()
+                .build_from_path(input_fn)
+                .with_context(|| format!("Error while opening {input_fn:?}:"))?;
+
+            // get the header
+            let mut vcf_header = vcf_reader.read_header()
+                .with_context(|| format!("Error while reading header of {input_fn:?}:"))?;
+
+            // update the header
+            vcf_header.insert("aardvark_version".parse()?, vcf::header::record::Value::from(cli_version.clone()))?;
+            vcf_header.insert("aardvark_command".parse()?, vcf::header::record::Value::from(cli_string.clone()))?;
+            for (header_key, header_value) in extra_header.iter() {
+                vcf_header.formats_mut().insert(header_key.clone(), header_value.clone());
+            }
+
+            let samples = vcf_header.sample_names_mut();
+            samples.clear();
+            samples.insert(sample_name);
+
+            let out_extension = match variant_source {
+                VariantSource::Truth => "truth.vcf.gz",
+                VariantSource::Query => "query.vcf.gz"
+            };
+
+            // make the VCF and write the header
+            let vcf_fn = debug_folder.join(out_extension);
+            debug!("Opening {vcf_fn:?} for writing...");
+
+            let file = File::create(&vcf_fn)?;
+            let w_threads = std::num::NonZeroUsize::new(threads.clamp(1, 4)).unwrap();
+            let bgzf_writer = bgzf::MultithreadedWriter::with_worker_count(w_threads, file);
+            let mut vcf_writer = vcf::io::Writer::new(bgzf_writer);
+
+            // write the header based on the data source
+            vcf_writer.write_header(&vcf_header)?;
+
             // save it to our writers list
-            assert!(vcf_writers.insert(source, vcf_writer).is_none());
+            let vc = Self {
+                source: variant_source,
+                filename: vcf_fn,
+                vcf_header,
+                vcf_writer
+            };
+            assert!(source_lookup.insert(variant_source, vc).is_none());
         }
 
-        Ok(Self {
-            truth_header: truth_vcf_header,
-            query_header: query_vcf_header,
-            vcf_writers
-        })
+        // now pop and return
+        let truth_writer = source_lookup.remove(&VariantSource::Truth).unwrap();
+        let query_writer = source_lookup.remove(&VariantSource::Query).unwrap();
+        Ok((truth_writer, query_writer))
     }
 
     /// Core result writer, which currently does not check the input order at all; so user beware.
@@ -142,17 +148,22 @@ impl VariantCategorizer {
     /// * `region` - the comparison region, which contains the loaded variants
     /// * `benchmark` - the results from our comparison, which contains the classification information
     pub fn write_results(&mut self, region: &CompareRegion, benchmark: &CompareBenchmark) -> anyhow::Result<()> {
-        // sanity checks
-        if region.truth_variants().len() != benchmark.truth_variant_data().len() {
-            bail!("Region and benchmark truth lengths are different for B#{}", region.region_id());
+        match self.source {
+            VariantSource::Truth => {
+                // sanity check then write it
+                if region.truth_variants().len() != benchmark.truth_variant_data().len() {
+                    bail!("Region and benchmark truth lengths are different for B#{}", region.region_id());
+                }
+                self.write_variants(VariantSource::Truth, region.region_id(), region.coordinates().chrom(), region.truth_variants(), region.truth_zygosity(), benchmark.truth_variant_data())?;
+            },
+            VariantSource::Query => {
+                // sanity check then write it
+                if region.query_variants().len() != benchmark.query_variant_data().len() {
+                    bail!("Region and benchmark query lengths are different for B#{}", region.region_id());
+                }
+                self.write_variants(VariantSource::Query, region.region_id(), region.coordinates().chrom(), region.query_variants(), region.query_zygosity(), benchmark.query_variant_data())?;
+            },
         }
-
-        if region.query_variants().len() != benchmark.query_variant_data().len() {
-            bail!("Region and benchmark query lengths are different for B#{}", region.region_id());
-        }
-
-        self.write_variants(VariantSource::Truth, region.region_id(), region.coordinates().chrom(), region.truth_variants(), region.truth_zygosity(), benchmark.truth_variant_data())?;
-        self.write_variants(VariantSource::Query, region.region_id(), region.coordinates().chrom(), region.query_variants(), region.query_zygosity(), benchmark.query_variant_data())?;
         Ok(())
     }
 
@@ -165,10 +176,8 @@ impl VariantCategorizer {
     /// * `zygosity` - the parsed zygosity
     /// * `bench_data` - the comparison data, which determines whether the variant is a TP, FN, or FP
     fn write_variants(&mut self, source: VariantSource, region_id: u64, chrom: &str, variants: &[Variant], zygosity: &[PhasedZygosity], bench_data: &[VariantMetrics]) -> anyhow::Result<()> {
-        let vcf_header = match source {
-            VariantSource::Truth => &self.truth_header,
-            VariantSource::Query => &self.query_header,
-        };
+        // make sure we didn't do something stupid
+        assert_eq!(self.source, source);
         for ((variant, &zygosity), bd) in variants.iter().zip(zygosity.iter()).zip(bench_data.iter()) {
             // convert REF and ALT back into strings
             let ref_allele = std::str::from_utf8(variant.allele0())?;
@@ -221,36 +230,33 @@ impl VariantCategorizer {
                 .build();
 
             // save the new records to the correct VCF file
-            let vcf_writer = self.vcf_writers.get_mut(&source).unwrap();
-            vcf_writer.write_variant_record(vcf_header, &record)?;
+            self.vcf_writer.write_variant_record(&self.vcf_header, &record)?;
         }
 
         Ok(())
+    }
+
+    pub fn filename(&self) -> &Path {
+        &self.filename
     }
 }
 
 /// Helpful utility function to index all outputs from a VariantCategorizer.
 /// Critically, it consumes the writer to finalize all outputs prior to indexing.
 /// # Arguments
-/// * `debug_folder` - the output folder we are indexing
 /// * `categorizer` - the writer that gets consumed
 /// # Errors
 /// * if the noodles indexing throws any errors; this could happen if the BED file provided is not sorted
-pub fn index_categorizer(debug_folder: &Path, categorizer: VariantCategorizer) -> anyhow::Result<()> {
+pub fn index_categorizer(categorizer: VariantCategorizer) -> anyhow::Result<()> {
+    // copy all the filenames, so we can index
+    let vcf_fn = categorizer.filename().to_path_buf();
+
     // drop it from memory, this forces all the finalizing to happen
     std::mem::drop(categorizer);
 
     // now index everything
-    let extensions = [
-        "truth.vcf.gz",
-        "query.vcf.gz",
-    ];
-
-    for extension in extensions.iter() {
-        let vcf_fn = debug_folder.join(extension);
-        debug!("Generating index for {vcf_fn:?}...");
-        crate::writers::noodles_idx::index_vcf(&vcf_fn)
-            .with_context(|| format!("Error while writing index for {vcf_fn:?}"))?;
-    }
+    debug!("Generating index for {vcf_fn:?}...");
+    crate::writers::noodles_idx::index_vcf(&vcf_fn)
+        .with_context(|| format!("Error while writing index for {vcf_fn:?}"))?;
     Ok(())
 }